+++ core/designsystem/src/main/kotlin/com/gradation/lift/designsystem/component/textField/LiftDefaultInputTextField.kt
@@ -1,0 +1,62 @@
+package com.gradation.lift.designsystem.component.textField
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import com.gradation.lift.designsystem.resource.LiftIcon
+import com.gradation.lift.designsystem.theme.LiftTheme
+
+
+@Composable
+fun LiftDefaultInputTextField(
+    modifier: Modifier = Modifier,
+    value: String,
+    placeHolderValue: String = "",
+    onValueChange: (String) -> Unit,
+    onValueClear: (() -> Unit)? = null,
+    enabled: Boolean = true,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
+    keyboardActions: KeyboardActions = KeyboardActions.Default,
+    isError: Boolean = false,
+    isValid: Boolean = false,
+
+    ) {
+
+    LiftBaseInputTextField(
+        modifier = modifier,
+        value = value,
+        placeHolderValue = placeHolderValue,
+        onValueChange = onValueChange,
+        keyboardOptions = keyboardOptions,
+        keyboardActions = keyboardActions,
+        isError = isError,
+        isValid = isValid,
+        enabled = enabled,
+        trailingIcon = @Composable {
+            if (isValid) {
+                Icon(
+                    modifier = modifier.size(LiftTheme.space.space24),
+                    painter = painterResource(id = LiftIcon.GreenCheck),
+                    contentDescription = "Valid",
+                    tint = Color.Unspecified
+                )
+            }
+            if (!isValid && value.isNotEmpty() && enabled && onValueClear != null) {
+                Icon(
+                    modifier = modifier
+                        .size(LiftTheme.space.space24)
+                        .clickable { onValueClear() },
+                    painter = painterResource(id = LiftIcon.Cancel),
+                    contentDescription = "Clear",
+                    tint = Color.Unspecified
+                )
+            }
+        }
+    )
+}
