+++ core/designsystem/src/main/kotlin/com/gradation/lift/designsystem/component/textField/LiftHeightWeightInputTextField.kt
@@ -1,0 +1,82 @@
+package com.gradation.lift.designsystem.component.textField
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import com.gradation.lift.designsystem.component.text.LiftText
+import com.gradation.lift.designsystem.component.text.LiftTextStyle
+import com.gradation.lift.designsystem.theme.LiftTheme
+
+
+@Composable
+fun LiftHeightInputTextField(
+    modifier: Modifier = Modifier,
+    value: String,
+    placeHolderValue: String = "",
+    onValueChange: (String) -> Unit,
+    enabled: Boolean = true,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
+    keyboardActions: KeyboardActions = KeyboardActions.Default,
+    isError: Boolean = false,
+    isValid: Boolean = false,
+) {
+    LiftBaseInputTextField(
+        modifier = modifier,
+        value = value,
+        placeHolderValue = placeHolderValue,
+        onValueChange = onValueChange,
+        keyboardOptions = keyboardOptions,
+        keyboardActions = keyboardActions,
+        isError = isError,
+        isValid = isValid,
+        enabled = enabled,
+        trailingIcon = @Composable {
+            LiftText(
+                modifier = modifier
+                    .padding(end = LiftTheme.space.space16),
+                textStyle = LiftTextStyle.No8,
+                text = "cm",
+                color = LiftTheme.colorScheme.no9,
+                textAlign = TextAlign.Start
+            )
+        }
+    )
+}
+
+@Composable
+fun LiftWeightInputTextField(
+    modifier: Modifier = Modifier,
+    value: String,
+    placeHolderValue: String = "",
+    onValueChange: (String) -> Unit,
+    enabled: Boolean = true,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
+    keyboardActions: KeyboardActions = KeyboardActions.Default,
+    isError: Boolean = false,
+    isValid: Boolean = false,
+) {
+    LiftBaseInputTextField(
+        modifier = modifier,
+        value = value,
+        placeHolderValue = placeHolderValue,
+        onValueChange = onValueChange,
+        keyboardOptions = keyboardOptions,
+        keyboardActions = keyboardActions,
+        isError = isError,
+        isValid = isValid,
+        enabled = enabled,
+        trailingIcon = @Composable {
+            LiftText(
+                modifier = modifier
+                    .padding(end = LiftTheme.space.space16),
+                textStyle = LiftTextStyle.No8,
+                text = "kg",
+                color = LiftTheme.colorScheme.no9,
+                textAlign = TextAlign.Start
+            )
+        }
+    )
+}
